+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0]
+- Supports course members completed and not completed count
+- Each rule can now it's own interval in days to check
+- Set checked rules and sent mails count in cron job status
+
 ## [1.1.0]
 - Supports course start and end field
 - Supports operator x days before or after